Remove export_name query
Part of #47320
