+++ src/librustc_typeck/diagnostics.rs
@@ -3774,6 +3774,29 @@ For more information about the inline attribute, https:
 read://doc.rust-lang.org/reference.html#inline-attributes
 "##,
 
+E0558: r##"
+The `export_name` attribute was malformed.
+
+Erroneous code example:
+
+```ignore (error-emitted-at-codegen-which-cannot-be-handled-by-compile_fail)
+#[export_name] // error: export_name attribute has invalid format
+pub fn something() {}
+
+fn main() {}
+```
+
+The `export_name` attribute expects a string in order to determine the name of
+the exported symbol. Example:
+
+```
+#[export_name = "some_function"] // ok!
+pub fn something() {}
+
+fn main() {}
+```
+"##,
+
 E0559: r##"
 An unknown field was specified into an enum's structure variant.
